requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
transformers>=4.35.2
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.1.0
numpy>=1.24.3
scikit-learn>=1.3.2
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
matplotlib>=3.8.2
